link to customer page

+++ application/api/services/databaseService.ts
@@ -29,7 +29,7 @@ export async function connectToTestMongoDatabase() {
     }
 
     mongod = await MongoMemoryServer.create();
-    mongoose.connect(mongod.getUri(), {});
+    await mongoose.connect(mongod.getUri(), {});
 }
 
 export async function closeDatabase() {
@@ -47,6 +47,6 @@ export async function clearDatabase() {
     const collections = mongoose.connection.collections;
     for (const key in collections) {
         const collection = collections[key];
-        await collection.deleteMany({});
+        collection && await collection.deleteMany({});
     }
 }
